# Use array for holding GOPATHs instead of map