```
$ go test -count=10 -run=NONE -bench=BenchmarkErrWithStackAndUTCTime -benchmem -cpuprofile=cpu-old.prof | tee old.log
goos: linux
goarch: amd64
pkg: github.com/hnakamur/ltsvlog
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17568 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17782 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17899 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     16502 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17250 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17588 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17489 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17762 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17544 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17680 ns/op	   16464 B/op	       4 allocs/op
PASS
ok  	github.com/hnakamur/ltsvlog	19.497s

$ go test -count=10 -run=NONE -bench=BenchmarkErrWithStackAndUTCTime -benchmem -cpuprofile=cpu-new.prof | tee new.log
goos: linux
goarch: amd64
pkg: github.com/hnakamur/ltsvlog
BenchmarkErrWithStackAndUTCTime-2   	  100000	     16743 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17368 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17271 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17686 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17489 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17307 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17460 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17110 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     16872 ns/op	   16464 B/op	       4 allocs/op
BenchmarkErrWithStackAndUTCTime-2   	  100000	     17201 ns/op	   16464 B/op	       4 allocs/op
PASS
ok  	github.com/hnakamur/ltsvlog	19.219s

$ benchstat old.log new.log | tee bench.log
name                      old time/op    new time/op    delta
ErrWithStackAndUTCTime-2    17.6µs ± 2%    17.3µs ± 3%  -2.08%  (p=0.005 n=9+10)

name                      old alloc/op   new alloc/op   delta
ErrWithStackAndUTCTime-2    16.5kB ± 0%    16.5kB ± 0%    ~     (all equal)

name                      old allocs/op  new allocs/op  delta
ErrWithStackAndUTCTime-2      4.00 ± 0%      4.00 ± 0%    ~     (all equal)
```